+++ backend-java/src/main/java/com/aixuniversity/maadictionary/dao/index/SearchFlatIndex.java
@@ -5,5 +5,5 @@ import it.unimi.dsi.fastutil.ints.IntArrayList;
 public interface SearchFlatIndex<F> extends SearchIndex<F>{
 
     IntArrayList allIds();
-    long totalFreq();
+    int totalFreq();
 }
